Refactoring: Move singature related functions from parser.py to sig_utils.py instead.
