+++ vyper/signatures/__init__.py
@@ -1,0 +1,5 @@
+from vyper.signatures.function_signature import (  # noqa: F401
+    FunctionSignature,
+    VariableRecord
+)
+from vyper.signatures.event_signature import EventSignature  # noqa: F401
